test: move to `integration` directory
