+++ packages/spectacular/src/lib/feature-testing/integration/angular-testing-library/docs-example.integration.spec.ts
@@ -9,10 +9,10 @@ import {
   crisisCenterPath,
   CrisisService,
 } from '@tour-of-heroes/crisis-center';
-import { SpectacularAppComponent } from '../../shared/app-component/spectacular-app.component';
-import { SpectacularFeatureTestingModule } from '../feature-testing-module/spectacular-feature-testing.module';
-import { SpectacularFeatureLocation } from '../navigation/spectacular-feature-location';
-import { SpectacularFeatureRouter } from '../navigation/spectacular-feature-router';
+import { SpectacularAppComponent } from '../../../shared/app-component/spectacular-app.component';
+import { SpectacularFeatureTestingModule } from '../../feature-testing-module/spectacular-feature-testing.module';
+import { SpectacularFeatureLocation } from '../../navigation/spectacular-feature-location';
+import { SpectacularFeatureRouter } from '../../navigation/spectacular-feature-router';
 
 describe('Tour of Heroes: Crisis center integration tests', () => {
   beforeEach(async () => {
