+++ packages/spectacular/src/lib/feature-testing/integration/angular-testing-library/edit-crisis/02-second-test.spec.ts
@@ -1,10 +1,12 @@
 import { render, screen } from '@testing-library/angular';
 import user from '@testing-library/user-event';
-import { CrisisCenterModule, crisisCenterPath } from '@tour-of-heroes/crisis-center';
-
-import { SpectacularAppComponent } from '../../../shared/app-component/spectacular-app.component';
-import { SpectacularFeatureTestingModule } from '../../feature-testing-module/spectacular-feature-testing.module';
-import { SpectacularFeatureRouter } from '../../navigation/spectacular-feature-router';
+import {
+  CrisisCenterModule,
+  crisisCenterPath,
+} from '@tour-of-heroes/crisis-center';
+import { SpectacularAppComponent } from '../../../../shared/app-component/spectacular-app.component';
+import { SpectacularFeatureTestingModule } from '../../../feature-testing-module/spectacular-feature-testing.module';
+import { SpectacularFeatureRouter } from '../../../navigation/spectacular-feature-router';
 
 it('Edit a crisis from crisis center home', async () => {
   const {
